--- Reverse-merging r2447 through r2424 into '.'
git-svn-id: svn://svn.code.sf.net/p/boomerang/code/trunk@2448 dc5e1367-85d4-4bd1-96b4-a20bfbfd92d5

+++ boomerang/db/exp.cpp
@@ -485,9 +485,9 @@ bool RefExp::operator==(const Exp& o) const
     if (((RefExp&)o).op != opSubscript) return false;
     if (!( *subExp1 == *((RefExp&)o).subExp1)) return false;
     // Allow a def of (Statement*)-1 as a wild card
-    if ((intptr_t)def == -1) return true;
+    if ((int)def == -1) return true;
     // Allow a def of NULL to match a def of an implicit assignment
-    if ((intptr_t)((RefExp&)o).def == -1) return true;
+    if ((int)((RefExp&)o).def == -1) return true;
     if (def == NULL && ((RefExp&)o).isImplicitDef()) return true;
     if (((RefExp&)o).def == NULL && def && def->isImplicit()) return true;
     return def == ((RefExp&)o).def;
@@ -1461,8 +1461,8 @@ void Exp::createDotFile(char* name)
 void Const::appendDotFile(std::ofstream& of)
 {
   // We define a unique name for each node as "e123456" if the address of "this" == 0x123456
-  of << "e" << std::hex << (intptr_t)this << " [shape=record,label=\"{";
-  of << operStrings[op] << "\\n0x" << std::hex << (intptr_t)this << " | ";
+  of << "e" << std::hex << (int)this << " [shape=record,label=\"{";
+  of << operStrings[op] << "\\n0x" << std::hex << (int)this << " | ";
   switch (op)
     {
     case opIntConst:
@@ -1489,13 +1489,13 @@ void Const::appendDotFile(std::ofstream& of)
 //	//	//	//
 void Terminal::appendDotFile(std::ofstream& of)
 {
-  of << "e" << std::hex << (intptr_t)this << " [shape=parallelogram,label=\"";
+  of << "e" << std::hex << (int)this << " [shape=parallelogram,label=\"";
   if (op == opWild)
     // Note: value is -1, so can't index array
     of << "WILD";
   else
     of << operStrings[op];
-  of << "\\n0x" << std::hex << (intptr_t)this;
+  of << "\\n0x" << std::hex << (int)this;
   of << "\"];\n";
 }
 
@@ -1505,9 +1505,9 @@ void Terminal::appendDotFile(std::ofstream& of)
 void Unary::appendDotFile(std::ofstream& of)
 {
   // First a node for this Unary object
-  of << "e" << std::hex << (intptr_t)this << " [shape=record,label=\"{";
-  // The (intptr_t) cast is to print the address, not the expression!
-  of << operStrings[op] << "\\n0x" << std::hex << (intptr_t)this << " | ";
+  of << "e" << std::hex << (int)this << " [shape=record,label=\"{";
+  // The (int) cast is to print the address, not the expression!
+  of << operStrings[op] << "\\n0x" << std::hex << (int)this << " | ";
   of << "<p1>";
   of << " }\"];\n";
 
@@ -1515,7 +1515,7 @@ void Unary::appendDotFile(std::ofstream& of)
   subExp1->appendDotFile(of);
 
   // Finally an edge for the subexpression
-  of << "e" << std::hex << (intptr_t)this << "->e" << (intptr_t)subExp1 << ";\n";
+  of << "e" << std::hex << (int)this << "->e" << (int)subExp1 << ";\n";
 }
 
 //	//	//	//
@@ -1524,15 +1524,15 @@ void Unary::appendDotFile(std::ofstream& of)
 void Binary::appendDotFile(std::ofstream& of)
 {
   // First a node for this Binary object
-  of << "e" << std::hex << (intptr_t)this << " [shape=record,label=\"{";
-  of << operStrings[op] << "\\n0x" << std::hex << (intptr_t)this << " | ";
+  of << "e" << std::hex << (int)this << " [shape=record,label=\"{";
+  of << operStrings[op] << "\\n0x" << std::hex << (int)this << " | ";
   of << "{<p1> | <p2>}";
   of << " }\"];\n";
   subExp1->appendDotFile(of);
   subExp2->appendDotFile(of);
   // Now an edge for each subexpression
-  of << "e" << std::hex << (intptr_t)this << ":p1->e" << (intptr_t)subExp1 << ";\n";
-  of << "e" << std::hex << (intptr_t)this << ":p2->e" << (intptr_t)subExp2 << ";\n";
+  of << "e" << std::hex << (int)this << ":p1->e" << (int)subExp1 << ";\n";
+  of << "e" << std::hex << (int)this << ":p2->e" << (int)subExp2 << ";\n";
 }
 
 //	//	//	//
@@ -1541,30 +1541,30 @@ void Binary::appendDotFile(std::ofstream& of)
 void Ternary::appendDotFile(std::ofstream& of)
 {
   // First a node for this Ternary object
-  of << "e" << std::hex << (intptr_t)this << " [shape=record,label=\"{";
-  of << operStrings[op] << "\\n0x" << std::hex << (intptr_t)this << " | ";
+  of << "e" << std::hex << (int)this << " [shape=record,label=\"{";
+  of << operStrings[op] << "\\n0x" << std::hex << (int)this << " | ";
   of << "{<p1> | <p2> | <p3>}";
   of << " }\"];\n";
   subExp1->appendDotFile(of);
   subExp2->appendDotFile(of);
   subExp3->appendDotFile(of);
   // Now an edge for each subexpression
-  of << "e" << std::hex << (intptr_t)this << ":p1->e" << (intptr_t)subExp1 << ";\n";
-  of << "e" << std::hex << (intptr_t)this << ":p2->e" << (intptr_t)subExp2 << ";\n";
-  of << "e" << std::hex << (intptr_t)this << ":p3->e" << (intptr_t)subExp3 << ";\n";
+  of << "e" << std::hex << (int)this << ":p1->e" << (int)subExp1 << ";\n";
+  of << "e" << std::hex << (int)this << ":p2->e" << (int)subExp2 << ";\n";
+  of << "e" << std::hex << (int)this << ":p3->e" << (int)subExp3 << ";\n";
 }
 //	//	//	//
 // TypedExp //
 //	//	//	//
 void TypedExp::appendDotFile(std::ofstream& of)
 {
-  of << "e" << std::hex << (intptr_t)this << " [shape=record,label=\"{";
-  of << "opTypedExp\\n0x" << std::hex << (intptr_t)this << " | ";
+  of << "e" << std::hex << (int)this << " [shape=record,label=\"{";
+  of << "opTypedExp\\n0x" << std::hex << (int)this << " | ";
   // Just display the C type for now
   of << type->getCtype() << " | <p1>";
   of << " }\"];\n";
   subExp1->appendDotFile(of);
-  of << "e" << std::hex << (intptr_t)this << ":p1->e" << (intptr_t)subExp1 << ";\n";
+  of << "e" << std::hex << (int)this << ":p1->e" << (int)subExp1 << ";\n";
 }
 
 //	//	//	//
@@ -1572,8 +1572,8 @@ void TypedExp::appendDotFile(std::ofstream& of)
 //	//	//	//
 void FlagDef::appendDotFile(std::ofstream& of)
 {
-  of << "e" << std::hex << (intptr_t)this << " [shape=record,label=\"{";
-  of << "opFlagDef \\n0x" << std::hex << (intptr_t)this << "| ";
+  of << "e" << std::hex << (int)this << " [shape=record,label=\"{";
+  of << "opFlagDef \\n0x" << std::hex << (int)this << "| ";
   // Display the RTL as "RTL <r1> <r2>..." vertically (curly brackets)
   of << "{ RTL ";
   int n = rtl->getNumStmt();
@@ -1581,7 +1581,7 @@ void FlagDef::appendDotFile(std::ofstream& of)
     of << "| <r" << std::dec << i << "> ";
   of << "} | <p1> }\"];\n";
   subExp1->appendDotFile(of);
-  of << "e" << std::hex << (intptr_t)this << ":p1->e" << (intptr_t)subExp1 << ";\n";
+  of << "e" << std::hex << (int)this << ":p1->e" << (int)subExp1 << ";\n";
 }
 
 /*==============================================================================
@@ -4537,7 +4537,7 @@ void Const::printx(int ind)
       std::cerr << u.pp->getName();
       break;
     default:
-      std::cerr << std::hex << "?" << (intptr_t)op << "?";
+      std::cerr << std::hex << "?" << (int)op << "?";
     }
   if (conscript)
     std::cerr << " \\" << std::dec << conscript << "\\";
@@ -4567,7 +4567,7 @@ void RefExp::printx(int ind)
   std::cerr << std::setw(ind) << " " << operStrings[op] << " ";
   std::cerr << "{";
   if (def == 0) std::cerr << "NULL";
-  else std::cerr << std::hex << (uintptr_t)def << "=" << std::dec << def->getNumber();
+  else std::cerr << std::hex << (unsigned)def << "=" << std::dec << def->getNumber();
   std::cerr << "}\n" << std::flush;
   child(subExp1, ind);
 }
