+++ boomerang/frontend/pentiumdecoder.h
@@ -37,13 +37,13 @@ public:
 	 * Decodes the machine instruction at pc and returns an RTL instance for
 	 * the instruction.
 	 */
-virtual DecodeResult& decodeInstruction (ADDRESS pc, intptr_t delta);
+virtual DecodeResult& decodeInstruction (ADDRESS pc, int delta);
 
 	/*
 	 * Disassembles the machine instruction at pc and returns the number of
 	 * bytes disassembled. Assembler output goes to global _assembly
 	 */
-virtual int decodeAssemblyInstruction (ADDRESS pc, intptr_t delta);
+virtual int decodeAssemblyInstruction (ADDRESS pc, int delta);
 
 private:
 	/*
@@ -61,7 +61,7 @@ private:
 	SWord	getWord(unsigned lc);
 	DWord	getDword(unsigned lc);
 
-	uintptr_t lastDwordLc;
+	unsigned lastDwordLc;
 };
 
 #endif
