+++ boomerang/frontend/ppcdecoder.h
@@ -37,13 +37,13 @@ public:
 	 * Decodes the machine instruction at pc and returns an RTL instance for
 	 * the instruction.
 	 */
-virtual DecodeResult& decodeInstruction (ADDRESS pc, intptr_t delta);
+virtual DecodeResult& decodeInstruction (ADDRESS pc, int delta);
 
 	/*
 	 * Disassembles the machine instruction at pc and returns the number of
 	 * bytes disassembled. Assembler output goes to global _assembly
 	 */
-virtual int			decodeAssemblyInstruction (ADDRESS pc, intptr_t delta);
+virtual int			decodeAssemblyInstruction (ADDRESS pc, int delta);
 
 
 private:
@@ -52,8 +52,8 @@ private:
 		 */
 		Exp*		dis_Eaddr(ADDRESS pc, int size = 0);
 		Exp*		dis_RegImm(ADDRESS pc);
-		Exp*		dis_Reg(unsigned int r);
-		Exp*		dis_RAmbz(unsigned int r);		// Special for rA of certain instructions
+		Exp*		dis_Reg(unsigned r);
+		Exp*		dis_RAmbz(unsigned r);		// Special for rA of certain instructions
 
 		void		unused(int x);
 		RTL*		createBranchRtl(ADDRESS pc, std::list<Statement*>* stmts, const char* name);
