+++ boomerang/loader/ElfBinaryFile.h
@@ -117,7 +117,7 @@ Elf32_Shdr;
 
 typedef struct {
     int st_name;
-    unsigned int st_value;
+    unsigned st_value;
     int st_size;
     unsigned char st_info;
     unsigned char st_other;
@@ -126,12 +126,12 @@ typedef struct {
 Elf32_Sym;
 
 typedef struct {
-    unsigned int r_offset;
+    unsigned r_offset;
     int r_info;
 }
 Elf32_Rel;
 
-#define	ELF32_R_SYM(info)		((info)>>8)
+#define	 ELF32_R_SYM(info)		 ((info)>>8)
 #define ELF32_ST_BIND(i)		((i) >> 4)
 #define ELF32_ST_TYPE(i)		((i) & 0xf)
 #define ELF32_ST_INFO(b, t)		(((b)<<4)+((t)&0xf))
@@ -184,9 +184,9 @@ public:
     //virtual ADDRESS	GetFirstHeaderAddress();		// Get ADDRESS of main header
     //		ADDRESS		GetNextHeaderAddress();			// Get any other headers
 
-    int8_t readNative1(ADDRESS a); // Read 1 bytes from native addr
-    int16_t readNative2(ADDRESS a); // Read 2 bytes from native addr
-    int32_t readNative4(ADDRESS a); // Read 4 bytes from native addr
+    int readNative1(ADDRESS a); // Read 1 bytes from native addr
+    int readNative2(ADDRESS a); // Read 2 bytes from native addr
+    int readNative4(ADDRESS a); // Read 4 bytes from native addr
     QWord readNative8(ADDRESS a); // Read 8 bytes from native addr
     float readNativeFloat4(ADDRESS a); // Read 4 bytes as float
     double readNativeFloat8(ADDRESS a); // Read 8 bytes as float
@@ -277,9 +277,9 @@ private:
     ADDRESS findRelPltOffset(int i, ADDRESS addrRelPlt, int sizeRelPlt, int numRelPlt, ADDRESS addrPlt);
 
     // Internal elf reading methods
-    int16_t elfRead2(int16_t* ps) const; // Read a short with endianness care
-    int32_t elfRead4(int32_t* pi) const; // Read an int with endianness care
-    void elfWrite4(int32_t* pi, int val); // Write an int with endianness care
+    int elfRead2(short* ps) const; // Read a short with endianness care
+    int elfRead4(int* pi) const; // Read an int with endianness care
+    void elfWrite4(int* pi, int val); // Write an int with endianness care
 
     FILE* m_fd; // File stream
     long m_lImageSize; // Size of image in bytes
