+++ boomerang/util/util.cpp
@@ -36,7 +36,6 @@
 #include <fstream>
 #include <cstdio>
 #include <cstring>
-#include <stdint.h>
 #include "util.h"
 
 #ifndef _WIN32
@@ -90,15 +89,15 @@ std::string initCapital(const std::string& s)
 bool hasExt(const std::string& s, const char* ext)
 {
   std::string tailStr = std::string(".") + std::string(ext);
-  uintptr_t i = s.rfind(tailStr);
+  unsigned int i = s.rfind(tailStr);
   if (i == std::string::npos)
     {
       return false;
     }
   else
     {
-      uintptr_t sLen = s.length();
-      uintptr_t tailStrLen = tailStr.length();
+      unsigned int sLen = s.length();
+      unsigned int tailStrLen = tailStr.length();
       return ((i + tailStrLen) == sLen);
     }
 }
@@ -137,7 +136,7 @@ std::string searchAndReplace( const std::string &in, const std::string &match,
                               const std::string &rep )
 {
   std::string result;
-  for ( intptr_t n = 0; n != -1; )
+  for ( int n = 0; n != -1; )
     {
       int l = in.find(match,n);
       result.append( in.substr(n,(l==-1?in.length() : l )-n) );
@@ -212,9 +211,9 @@ void escapeXMLChars(std::string &s)
   const char *replace[] =
     { "&lt;", "&gt;", "&amp;"
     };
-  for (uintptr_t i = 0; i < s.size(); i++)
+  for (unsigned i = 0; i < s.size(); i++)
     {
-      uintptr_t n = bad.find(s[i]);
+      unsigned n = bad.find(s[i]);
       if (n != std::string::npos)
         {
           s.replace(i, 1, replace[n]);
@@ -254,7 +253,7 @@ char* escapeStr( const char* str )
           if (!escapedSucessfully)
             {
               // it isn't so just use the \xhh escape
-              out << "\\x" << std::hex << std::setfill('0') << std::setw(2) << (intptr_t)*str;
+              out << "\\x" << std::hex << std::setfill('0') << std::setw(2) << (int)*str;
               out << std::setfill(' ');
             }
         }
